Update models for Backslot changes
This updates all of the models that are affected by the changes I'm proposing for BackSlot & BackSlotAddon; where we use the "head" category for scaling and positioning rather than the backslot's original usage of the "ground" category.
This allows for items to be orientated for backslot on a per item basis and without affecting other parts of the game as pretty much all items that would be used in Backslot do not go into the "head" slot, so it's otherwise unused by these items.
Most files only include the addition of the "head" category or changes within it, however there were a few files that required rotating the base model (and therefore requiring updates to third person, first person, etc.) but they should functionally be exactly the same as before these changes within the game.